Rename fromConfig to get
